Recipe App: Recipe Details

+++ db/migrate/20230313092040_create_recipes.rb
@@ -2,8 +2,8 @@ class CreateRecipes < ActiveRecord::Migration[7.0]
   def change
     create_table :recipes do |t|
       t.string :name
-      t.decimal :preparation_time
-      t.decimal :cooking_time
+      t.float :preparation_time
+      t.float :cooking_time
       t.text :description
       t.boolean :public, default: false
       t.references :user, null: false, foreign_key: true, index: true
